Docker conflict fox from merge to master

+++ docker/build/build.sh
@@ -2,7 +2,7 @@
 echo "Build script started"
 
 rm -rf /output/*
-rm -rf /output-docker-image/*
+
 
 cd /srcgmd/
 ls -lrth
@@ -22,16 +22,21 @@ echo 'jlink --add-modules "java.base,java.compiler,java.desktop,java.instrument,
 echo '#!/bin/bash' > /output/start.sh
 echo 'jre-for-gmd/bin/java -jar CoopNetwork.jar -cp lib\* -cp conf -Dnxt.runtime.dirProvider=nxt.env.DefaultDirProvider nxt.Nxt' >> /output/start.sh
 chmod +x /output/start.sh
+chmod -R oug+w /output
 
-cp /docker-image/Dockerfile /output/
-cd /output/
-docker build -t "gasiminei/gmd-node" .
-rm -f /output/Dockerfile
-docker image save -o /output-docker-image/gmd-node.tar gmd-node
-cp -f /docker-image/docker-compose.yml /output-docker-image/docker-compose.yml
-gzip /output-docker-image/gmd-node.tar
-mkdir -p /output-docker-image/conf
-touch /output-docker-image/conf/nxt.properties
-
+if [[ $1 && $1 = "dockerimg" ]]; then 
+	rm -rf /output-docker-image/*
+	cp /docker-image/Dockerfile /output/
+	cd /output/
+	docker build -t "coopnetwork/gmd-node" .
+	rm -f /output/Dockerfile
+	docker image save -o /output-docker-image/gmd-node.tar coopnetwork/gmd-node
+	cp -f /docker-image/docker-compose.yml /output-docker-image/docker-compose.yml
+	cp -f /docker-image/docker-compose-rproxy.yml /output-docker-image/docker-compose-rproxy.yml
+	#gzip /output-docker-image/gmd-node.tar
+	mkdir -p /output-docker-image/conf
+	chmod -R oug+w /output-docker-image
+	touch /output-docker-image/conf/nxt.properties
+fi
 
 echo "Build script finished"
